+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### Version - 0.9.20.0
 * Don't reuse HTTP request objects (#532)
 * Block popups in the in-app browser (#535)
+* Don't print API keys in logs (#532)
+* Store xxHash caches in binary format (#530)
 
 
 #### Version - 0.9.19.0